unify metal.py and redfish.py as the UUID changes based on system depth
